pyproject.toml
[project]
name = "webscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "typer>=0.15.2",
]

[dependency-groups]
dev = [
    "fastapi>=0.115.6",
    "coverage>=7.8.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pylint>=3.3.6",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.3",
]
